Updates to History Strategy
Attempting to update how the history data is pulled.

When battlesGetData.js runs, it creates a dated history file in the history folder.

When battlesGet.js runs, it creates individual player history files.

I also created a mergeHistory.js that merges all the files in the history folder

And a mergePlayerHistory.js that merges the player histories into one file that is put into the history folder.

I have something wrong here as the data isn't merging correctly - but the goal is the deal with the fact that pulling history only pulls the last 50 battles at a time. By creating dated files for the pull and merging them, we can create a more exhaustive history set.